requirements.txt
Pillow
requests
bleak
qrcode[pil]